requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
vaderSentiment>=3.3.2
pyttsx3>=2.90
edge-tts>=7.0.0
